+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,4 +1,7 @@
-﻿
+﻿### Version 3.5.1
+
+- Improvement: Added support for `Value Object` collection mapping.
+
 ### Version 3.5.0
 
 - Improvement: `IHasCSharpStatements`'s `FindStatement` method now returns the child statement match (the more specific match) rather than parent match if there are multiple matches.